common_modules/rag/pyproject.toml
[project]
name = "rag-ingest"
version = "0.1.0"
description = "RAG ingestion script for Vertex Matching Engine"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-cloud-aiplatform==1.67.0",
    "google-auth==2.36.0",
    "vertexai>=1.38.0",
]

[tool.uv]
package = false

